# visitor-register-main/lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Visitor {
  id: number;
  name: string;
  phone: string;
  personToMeet: string;
  purpose: string;
  checkedInAt: string;
}

export interface VisitorInput {
  /** @minLength 1 */
  name: string;
  /** @minLength 1 */
  phone: string;
  /** @minLength 1 */
  personToMeet: string;
  /** @minLength 1 */
  purpose: string;
}

export interface VisitorSummary {
  total: number;
  today: number;
  /** @nullable */
  latestCheckIn: string | null;
}

export type ListVisitorsParams = {
search?: string;
};