# visitor-register-main/lib/api-zod/src/generated/types/visitor.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Visitor {
  id: number;
  name: string;
  phone: string;
  personToMeet: string;
  purpose: string;
  checkedInAt: Date;
}