# visitor-register-main/lib/api-zod/src/generated/types/visitorInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface VisitorInput {
  /** @minLength 1 */
  name: string;
  /** @minLength 1 */
  phone: string;
  /** @minLength 1 */
  personToMeet: string;
  /** @minLength 1 */
  purpose: string;
}